Preview string beginnings and ends

Use str_head() to view the beginning of a string and str_tail() to view the end of a string. These functions work better for longer strings.

Usage

str_head(string, width = NULL, ellipsis = "...")

str_tail(string, width = NULL, ellipsis = "...")

Arguments

string

A character vector of strings to preview.

width

A single integer for the number of characters to show, without consideration for the length of ellipsis. If not supplied, defaults to the length of the longest string in string divided by 10 and rounded up to the nearest integer.

ellipsis

A single string used to indicate that content has been removed.

Value

A character vector the same length as string.

See also

Examples

# Basic usage
str_head(paste0(rep(letters, 2), collapse=""))
#> [1] "abc..."
str_tail(paste0(rep(letters, 2), collapse=""))
#> [1] "...xyz"

# Slightly more realistic usage
library(tibble)
library(dplyr, warn.conflicts=FALSE)
library(stringi)

my_docs <- tibble(
  doc_id = stri_rand_strings(4, length=6),
  text = rep(4, 4)
) %>%
mutate(text = stri_rand_lipsum(text, start_lipsum=FALSE))

my_docs %>%
  pull(text, name=doc_id) %>%
  str_head()
#>                                                                                            5pb90S 
#> "In conubia, ut hac maximus amet conubia sed volutpat vitae. Class cursus elit, mauris porta...." 
#>                                                                                            UHjlsA 
#> "Ornare mi lectus vel sed penatibus amet mollis. Hendrerit vehicula. Lobortis class sed eu et..." 
#>                                                                                            2JOCP3 
#> "Odio potenti blandit convallis volutpat mollis erat integer laoreet. Aliquam ac et ultricies..." 
#>                                                                                            OyHgjC 
#> "Finibus rhoncus, erat in dictum urna nulla. Quis tempus vel malesuada dictum nunc fusce inte..." 

my_docs %>%
  pull(text, name=doc_id) %>%
  str_tail()
#>                                                                                            5pb90S 
#> "...mpus et, nisl amet commodo, phasellus. Cum vel parturient nisl tristique viverra ut, tortor." 
#>                                                                                            UHjlsA 
#> "...risque at sed eu eleifend maecenas ipsum. Nisl, aenean efficitur sociosqu lorem blandit nec." 
#>                                                                                            2JOCP3 
#> "...stibulum per tellus elementum quis aliquam donec finibus sed in quis volutpat vestibulum et." 
#>                                                                                            OyHgjC 
#> "...tor blandit laoreet tortor neque. Sed mus lacus facilisi tempus egestas tempor ut penatibus."